Add Scoped file and add marks
